# improve performance of reduction on outer axis